fix previous commit -> was something wrong with Sourcetree, previous commit save not all modification.

+++ Internship/iOS/2018-KR/Converter/Categories/NSString+Formatting.h
@@ -10,7 +10,7 @@
 
 @interface NSString (Formatting)
 
-+ (NSString *)TYstringWithInt:(NSInteger)number;
++ (NSString *)TYstringWithInt:(long long)number;
 
 + (NSString *)TYstringWithLeadingWhitespace:(NSString *)string;
