+++ Internship/iOS/2018-KR/Converter/Formatter/UkrainianFormatter.m
@@ -47,7 +47,7 @@ static NSString * kGenitiveForm     = @"genitiveForm";
 }
 
 //  1..9
-- (NSString *)unitsFormatter:(NSInteger)number multiplier:(NSInteger)multiplier {
+- (NSString *)unitsFormatter:(NSInteger)number multiplier:(long long)multiplier {
     NSString *unitsString = nil;
     NSString *key = [NSString TYstringWithInt:number];
     
@@ -60,19 +60,19 @@ static NSString * kGenitiveForm     = @"genitiveForm";
     return [NSString TYstringWithLeadingWhitespace:unitsString];
 }
 //  10..19
-- (NSString *)teensFormatter:(NSInteger)number multiplier:(NSInteger)multiplier {
+- (NSString *)teensFormatter:(NSInteger)number multiplier:(long long)multiplier {
     NSString *teensString = self.numerals.cardinal[[NSString TYstringWithInt:number]];
     
     return [NSString TYstringWithLeadingWhitespace:teensString];
 }
 //  20, 30, 40..90
-- (NSString *)roundTensFormatter:(NSInteger)number multiplier:(NSInteger)multiplier {
+- (NSString *)roundTensFormatter:(NSInteger)number multiplier:(long long)multiplier {
     NSString *roundTensString = self.numerals.cardinal[[NSString TYstringWithInt:number]];
     
     return [NSString TYstringWithLeadingWhitespace:roundTensString];
 }
 //  21, 22, 23..99
-- (NSString *)tensFormatter:(NSInteger)number multiplier:(NSInteger)multiplier {
+- (NSString *)tensFormatter:(NSInteger)number multiplier:(long long)multiplier {
     NSInteger units = number % 10;
     NSInteger roundTens = number - units;
     
@@ -84,14 +84,14 @@ static NSString * kGenitiveForm     = @"genitiveForm";
     return [@[roundTensString, unitsString] componentsJoinedByString:kWHITESPACE];
 }
 //  100, 200, 300..900
-- (NSString *)hundredsFormatter:(NSInteger)number multiplier:(NSInteger)multiplier {
+- (NSString *)hundredsFormatter:(NSInteger)number multiplier:(long long)multiplier {
     NSString *hundredsString = self.numerals.cardinal[[NSString TYstringWithInt:number]];
     
     return [NSString TYstringWithLeadingWhitespace:hundredsString];
 }
 
 
-- (NSString *)largeNumbersFormatter:(NSInteger)multiplier quantity:(NSInteger)quantity {
+- (NSString *)largeNumbersFormatter:(long long)multiplier quantity:(NSInteger)quantity {
     NSString *result = kEMPTY_STRING;
     NSString *key = [NSString TYstringWithInt:multiplier];
     NSInteger tens  = quantity % 100;
@@ -126,7 +126,7 @@ static NSString * kGenitiveForm     = @"genitiveForm";
 }
 
 //  MARK:  common formatters
-- (NSString *)starterFormatter:(NSInteger)number {
+- (NSString *)starterFormatter:(long long)number {
     NSString *result = nil;
     NSString *key = [NSString TYstringWithInt:number];
     
@@ -199,7 +199,7 @@ static NSString * kGenitiveForm     = @"genitiveForm";
 
 
 
-- (NSString *)finishingFormatter:(NSInteger)number withString:(NSString *)string {
+- (NSString *)finishingFormatter:(long long)number withString:(NSString *)string {
     return [string TYstringByTrimmingWhitespace];
 };
